[FEAT] 뷰모델 추가링
추가링추가링
