+++ lib/user/home/main_page/mogulist_page/widgets/mogu_history_card.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import '../../../post/post_detail_page.dart';
+import '../../../post/post_detail_page/post_detail_page.dart';
 
 class MoguHistoryCard extends StatefulWidget {
   final Map<String, dynamic> post;
