+++ lib/user/home/main_page/my_page/my_page.dart
@@ -1,15 +1,10 @@
-import 'dart:convert';
 import 'package:flutter/material.dart';
-import 'package:flutter_dotenv/flutter_dotenv.dart';
-import 'package:http/http.dart' as http;
-import 'package:intl/intl.dart';
-import 'package:mogu_app/user/home/main_page/common/common_methods.dart';
-import 'package:shared_preferences/shared_preferences.dart';
+import 'package:mogu_app/user/home/main_page/my_page/my_page_viewModel.dart';
+import 'package:provider/provider.dart';
 
-import '../../../../service/location_service.dart';
-import '../../../myPage/account_management_page.dart';
-import '../../../myPage/setting_page.dart';
-import '../../../myPage/update_profile_page.dart';
+import '../../../myPage/account_management_page/account_management_page.dart';
+import '../../../myPage/setting_page/setting_page.dart';
+import '../../../myPage/update_profile_page/update_profile_page.dart';
 import '../bottom/home_page_bottom.dart';
 
 class MyPage extends StatefulWidget {
@@ -22,184 +17,17 @@ class MyPage extends StatefulWidget {
 }
 
 class _MyPageState extends State<MyPage> {
-  late Map<String, dynamic> userInfo;
-  late String userId;
-  late String name;
-  late String nickname;
-  late String phone;
-  late int level;
-  late String manner;
-  late double longitude;
-  late double latitude;
-  late int distanceMeters;
-  late DateTime registerDate;
-  late String profileImage;
-  late String address = "Loading location...";
-  late String token;
-
-  int currentPurchaseCount = 0;
-  int needPurchaseCount = 0;
-  int savingCost = 0;
 
   @override
   void initState() {
     super.initState();
-    _getUserInfo().then((value) {
-      _initializeUserInfo();
-      _getAddress();
-      findUserLevel(context);
-      findUserSavingCost(context);
-    });
-  }
-
-  Future<void> _getUserInfo() async {
-    SharedPreferences pref = await SharedPreferences.getInstance();
-    String? userJson = pref.getString('userJson');
-    userJson == null ? showDialog(
-      context: context,
-      builder: (context) {
-        return AlertDialog(
-          title: Text('로그인 정보가 전달되지 않았습니다.'),
-        );
-      }
-    ) :
-    setState(() {
-      userInfo = jsonDecode(userJson);
-    });
-  }
-
-  void _initializeUserInfo() {
-    userId = userInfo['userId'] ?? '';
-    token = userInfo['token'] ?? '';
-    name = userInfo['name'] ?? '';
-    nickname = userInfo['nickname'] ?? '';
-    phone = userInfo['phone'] ?? '';
-    level = userInfo['level'] ?? 0;
-    manner = userInfo['manner'] ?? '';
-    longitude = userInfo['longitude']?.toDouble() ?? 0.0;
-    latitude = userInfo['latitude']?.toDouble() ?? 0.0;
-    distanceMeters = userInfo['distanceMeters'] ?? 0;
-    registerDate = DateTime.parse(userInfo['registerDate'] ?? DateTime.now().toIso8601String());
-    profileImage = userInfo['profileImage'] ?? '';
-  }
-
-  Future<void> _getUpdatedUserInfo() async {
-    final updatedUserInfo = await fetchUpdatedUserInfo(userId, token);
-    if (updatedUserInfo != null) {
-      setState(() {
-        nickname = updatedUserInfo['nickname'] ?? nickname;
-        profileImage = updatedUserInfo['profileImage'] ?? profileImage;
-        longitude = updatedUserInfo['longitude']?.toDouble() ?? longitude;
-        latitude = updatedUserInfo['latitude']?.toDouble() ?? latitude;
-      });
-      await _getAddress();
-    } else {
-      CommonMethods.showErrorDialog(context, '오류', '사용자 정보를 불러올 수 없습니다.');
-    }
-  }
-
-  Future<Map<String, dynamic>?> fetchUpdatedUserInfo(String userId, String token) async {
-    String url = 'http://${dotenv.env['SERVER_IP']}:${dotenv.env['SERVER_PORT']}/user/my';
-
-    try {
-      final response = await http.get(
-        Uri.parse(url),
-        headers: {
-          'Authorization': token,
-          'Content-Type': 'application/json; charset=UTF-8',
-        },
-      );
-
-      if (response.statusCode == 200) {
-        return jsonDecode(response.body);
-      } else if (response.statusCode == 401) {
-        CommonMethods.showErrorDialog(context, '오류', '인증 오류: 로그인 정보가 유효하지 않습니다.');
-        return null;
-      } else {
-        CommonMethods.showErrorDialog(context, '오류', '사용자 정보를 불러올 수 없습니다.');
-        return null;
-      }
-    } catch (e) {
-      CommonMethods.showErrorDialog(context, '오류', '서버와의 연결 오류가 발생했습니다: ${e.toString()}');
-      return null;
-    }
-  }
-
-  Future<void> _getAddress() async {
-    try {
-      String fetchedAddress = await LocationService().getAddressFromCoordinates(latitude, longitude);
-      setState(() {
-        address = fetchedAddress;
-      });
-    } catch (e) {
-      setState(() {
-        address = "주소를 가져올 수 없습니다.";
-      });
-    }
-  }
-
-  Future<void> findUserLevel(BuildContext context) async {
-    String url = 'http://${dotenv.env['SERVER_IP']}:${dotenv.env['SERVER_PORT']}/user/level';
-
-    try {
-      final response = await http.get(
-        Uri.parse(url),
-        headers: <String, String>{
-          'Authorization': token,
-          'Content-Type': 'application/json; charset=UTF-8',
-        },
-      );
-
-      if (response.statusCode == 200) {
-        final Map<String, dynamic> data = jsonDecode(response.body);
-
-        setState(() {
-          // userUid = data['userUid']; // 여기선 uid가 필요가 없어유
-          level = data['level'];
-          currentPurchaseCount = data['currentPurchaseCount'];
-          needPurchaseCount = data['needPurchaseCount'];
-        });
-      } else {
-        CommonMethods.showErrorDialog(context, '오류', '서버에서 오류가 발생했습니다.');
-      }
-    } catch (e) {
-      CommonMethods.showErrorDialog(context, '오류', '서버와의 연결 오류가 발생했습니다.');
-    }
-  }
-
-  Future<void> findUserSavingCost(BuildContext context) async {
-    String url = 'http://${dotenv.env['SERVER_IP']}:${dotenv.env['SERVER_PORT']}/user/saving';
-
-    try {
-      final response = await http.get(
-        Uri.parse(url),
-        headers: <String, String>{
-          'Authorization': token,
-          'Content-Type': 'application/json; charset=UTF-8',
-        },
-      );
-
-      if (response.statusCode == 200) {
-        final Map<String, dynamic> data = jsonDecode(response.body);
-
-        setState(() {
-          savingCost = data['savingCost'];
-        });
-      } else {
-        CommonMethods.showErrorDialog(context, '오류', '서버에서 오류가 발생했습니다.');
-      }
-    } catch (e) {
-      CommonMethods.showErrorDialog(context, '오류', '서버와의 연결 오류가 발생했습니다.');
-    }
-  }
-
-  String formatCurrency(int amount) {
-    final formatter = NumberFormat('#,###');
-    return formatter.format(amount);
+    final viewModel = Provider.of<MyPageViewModel>(context, listen: false);
+    viewModel.initViewModel(context);
   }
 
   @override
   Widget build(BuildContext context) {
+    final viewModel = Provider.of<MyPageViewModel>(context);
     return Scaffold(
       appBar: AppBar(
         title: Text(
@@ -238,9 +66,10 @@ class _MyPageState extends State<MyPage> {
                             Navigator.push(
                               context,
                               MaterialPageRoute(
-                                builder: (context) => AccountManagementPage(userInfo: userInfo),
+                                builder: (context) =>
+                                  AccountManagementPage(userInfo: viewModel.userInfo),
                               ),
-                            ).then((_) => _getUpdatedUserInfo());
+                            ).then((_) => viewModel.getUpdatedUserInfo(context));
                           },
                         ),
                         ListTile(
@@ -289,10 +118,10 @@ class _MyPageState extends State<MyPage> {
                       child: CircleAvatar(
                         radius: 40,
                         backgroundColor: Colors.grey.shade300,
-                        backgroundImage: profileImage.isNotEmpty
-                            ? NetworkImage(profileImage)
+                        backgroundImage: viewModel.profileImage.isNotEmpty
+                            ? NetworkImage(viewModel.profileImage)
                             : null,
-                        child: profileImage.isEmpty
+                        child: viewModel.profileImage.isEmpty
                             ? Icon(Icons.person, size: 50, color: Colors.white)
                             : null,
                       ),
@@ -304,7 +133,7 @@ class _MyPageState extends State<MyPage> {
                       textBaseline: TextBaseline.alphabetic,
                       children: [
                         Text(
-                          'Lv.$level',
+                          'Lv.${viewModel.level}',
                           style: TextStyle(
                             fontSize: 16,
                             fontWeight: FontWeight.bold,
@@ -313,7 +142,7 @@ class _MyPageState extends State<MyPage> {
                         ),
                         SizedBox(width: 15), // 간격을 자연스럽게 조절
                         Text(
-                          nickname,
+                          viewModel.nickname,
                           style: TextStyle(
                             fontSize: 24, // 닉네임을 강조하기 위해 더 크게 설정
                             fontWeight: FontWeight.bold,
@@ -327,18 +156,18 @@ class _MyPageState extends State<MyPage> {
                           context,
                           MaterialPageRoute(
                             builder: (context) => UpdateProfilePage(
-                              userId: userId,
-                              nickname: nickname,
-                              profileImage: profileImage,
-                              longitude: longitude,
-                              latitude: latitude,
-                              token: token,
+                              userId: viewModel.userId,
+                              nickname: viewModel.nickname,
+                              profileImage: viewModel.profileImage,
+                              longitude: viewModel.longitude,
+                              latitude: viewModel.latitude,
+                              token: viewModel.token,
                             ),
                           ),
                         );
 
                         if (updatedUserInfo != null) {
-                          await _getUpdatedUserInfo(); // 프로필 수정 후 사용자 정보를 다시 가져옵니다.
+                          await viewModel.getUpdatedUserInfo(context); // 프로필 수정 후 사용자 정보를 다시 가져옵니다.
                         }
                       },
                       style: ElevatedButton.styleFrom(
@@ -358,12 +187,13 @@ class _MyPageState extends State<MyPage> {
                             Text('현재 거래 횟수', style: TextStyle(fontSize: 16, color: Color(0xFFB34FD1))),
                             SizedBox(height: 10),
                             Text(
-                              '$currentPurchaseCount / $needPurchaseCount',
+                              '${viewModel.currentPurchaseCount} / ${viewModel.needPurchaseCount}',
                               style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
                             ),
                             SizedBox(height: 4),
                             Text(
-                              '다음 레벨까지 ${(needPurchaseCount - currentPurchaseCount).abs()}번 남음',
+                              '다음 레벨까지 ${(viewModel.needPurchaseCount -
+                                  viewModel.currentPurchaseCount).abs()}번 남음',
                               style: TextStyle(fontSize: 14, color: Colors.grey),
                             ),
                           ],
@@ -373,7 +203,7 @@ class _MyPageState extends State<MyPage> {
                             Text('매너도', style: TextStyle(fontSize: 16, color: Color(0xFFB34FD1))),
                             SizedBox(height: 10),
                             Text(
-                              manner,
+                              viewModel.manner,
                               style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
                             ),
                           ],
@@ -384,19 +214,19 @@ class _MyPageState extends State<MyPage> {
                     ListTile(
                       leading: Icon(Icons.location_on, color: Color(0xFFB34FD1)),
                       title: Text('나의 위치', style: TextStyle(color: Color(0xFFB34FD1))),
-                      subtitle: Text(address),
+                      subtitle: Text(viewModel.address),
                     ),
                     ListTile(
                       leading: Icon(Icons.calendar_today, color: Color(0xFFB34FD1)),
                       title: Text('가입일', style: TextStyle(color: Color(0xFFB34FD1))),
-                      subtitle: Text('${registerDate.year}/${registerDate.month}/${registerDate.day}',
+                      subtitle: Text('${viewModel.registerDate.year}/${viewModel.registerDate.month}/${viewModel.registerDate.day}',
                           style: TextStyle(fontSize: 18)),
                     ),
                     ListTile(
                       leading: Icon(Icons.money, color: Color(0xFFB34FD1)),
                       title: Text('모구로 아낌비용', style: TextStyle(color: Color(0xFFB34FD1))),
                       subtitle: Text(
-                        '${formatCurrency(savingCost)}원',
+                        '${viewModel.formatCurrency(viewModel.savingCost)}원',
                         style: TextStyle(fontSize: 18),
                       ),
                     ),
