+++ lib/user/myPage/update_profile_page/update_profile_page.dart
@@ -6,7 +6,7 @@ import 'package:image_picker/image_picker.dart';
 import 'package:http/http.dart' as http;
 import 'package:http_parser/http_parser.dart';
 
-import '../../service/location_service.dart';
+import '../../../service/location_service.dart';
 
 class UpdateProfilePage extends StatefulWidget {
   final String userId;
